backend/requirements.txt
accelerate==1.1.1
adal==1.2.7
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
argh==0.31.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.5
asyncer==0.0.7
attrs==23.2.0
Authlib==1.3.1
# MCP (Model Context Protocol) dependencies
mcp>=1.0.0
autogen-agentchat==0.2.36
autopep8==2.3.1
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.17.1
azure-ai-projects==1.0.0b8
azure-ai-vision==0.8.0a1
azure-cognitiveservices-speech==1.37.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.7.0
azure-functions==1.21.3
azure-graphrbac==0.61.2
azure-identity==1.17.1
azure-keyvault-secrets==4.9.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.0.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-monitor-opentelemetry-exporter==1.0.0b27
azure-search-documents==11.5.2
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-storage-file-share==12.16.0
azureml==0.2.7
azureml-core==1.58.0
babel==2.17.0
backoff==2.2.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chainlit==1.3.0rc1
chardet==5.2.0
charset-normalizer==3.3.2
chess==1.10.0
chevron==0.14.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
email_validator==2.2.0
emoji==2.14.0
eval_type_backport==0.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.15.4
filetype==1.2.0
fixedint==0.1.6
FLAML==2.1.2
Flask==3.0.3
Flask-Cors==4.0.1
flask-restplus==0.13.0
flask-restx==1.3.0
fonttools==4.53.1
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.32.0
google_search_results==2.4.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
griffe==1.6.0
grpcio==1.68.0
grpcio-health-checking==1.68.0
grpcio-tools==1.68.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==24.3.1
kiwisolver==1.4.5
knack==0.12.0
langchain==0.3.7
langchain-community==0.3.2
langchain-core==0.3.33
langchain-openai==0.2.9
langchain-text-splitters==0.3.0
langdetect==1.0.9
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.135
Lazify==0.4.0
literalai==0.0.623
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.3.0
moviepy==2.1.2
mpmath==1.3.0
msal==1.29.0
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
mysql-connector-python==9.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.18
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
openai==1.70.0
openai-agents==0.0.3
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opencv-python==4.11.0.86
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandas-datareader==0.10.0
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pathspec==0.12.1
peewee==3.17.6
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.2.2
portalocker==2.10.0
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.47
promptflow==1.13.0
promptflow-azure==1.13.0
promptflow-core==1.13.0
promptflow-devkit==1.13.0
promptflow-tools==1.4.0
promptflow-tracing==1.13.0
prompty==0.1.24
proto-plus==1.24.0
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.32
pycodestyle==2.12.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.5.2
pydantic_core==2.27.2
pydash==7.0.7
PyDispatcher==2.0.7
Pygments==2.18.0
PyJWT==2.8.0
pymssql==2.3.2
PyNaCl==1.5.0
pynndescent==0.5.13
pyodbc==5.2.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pypdf==5.0.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-iso639==2024.4.27
python-json-logger==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-socketio==5.11.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.10.0
readme_renderer==44.0
referencing==0.35.1
regex==2024.5.15
replicate==0.34.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==71.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
strictyaml==1.7.3
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.20.1
tomli==2.2.1
tool==0.8.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.0
twine==5.1.1
typer==0.12.3
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.7
unstructured==0.15.14
unstructured-client==0.26.1
uptrace==1.28.2
uri-template==1.3.0
urllib3==2.2.2
utils==1.0.2
uvicorn==0.25.0
validators==0.34.0
waitress==2.1.2
watchfiles==0.20.0
wcwidth==0.2.13
weaviate-client==4.9.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.2
